+++ src/main/java/de/hysky/skyblocker/skyblock/item/background/ItemBackgroundManager.java
@@ -3,7 +3,7 @@ package de.hysky.skyblocker.skyblock.item.background;
 import de.hysky.skyblocker.annotations.Init;
 import de.hysky.skyblocker.skyblock.item.background.adders.ItemRarityBackground;
 import de.hysky.skyblocker.skyblock.item.background.adders.JacobMedalBackground;
-import de.hysky.skyblocker.skyblock.item.background.adders.AttributeBackground;
+import de.hysky.skyblocker.skyblock.item.background.adders.LegacyAttributeBackground;
 import de.hysky.skyblocker.utils.scheduler.Scheduler;
 import net.fabricmc.fabric.api.client.screen.v1.ScreenEvents;
 import net.minecraft.client.gui.DrawContext;
@@ -16,7 +16,7 @@ public class ItemBackgroundManager {
 	private static final List<ColoredItemBackground<?>> BACKGROUNDS = List.of(
 			new ItemRarityBackground(),
 			new JacobMedalBackground(),
-			new AttributeBackground()
+			new LegacyAttributeBackground()
 	);
 
 	@Init
